# perf: reduce tst built time