```js
import { bench, group, run } from 'mitata'
import {
  wellknownHeaderNames,
  headerNameLowerCasedRecord
} from '../lib/core/constants.js'
import { TernarySearchTree } from '../lib/core/tree.js'
// copy new implementation
import { TernarySearchTree2 } from '../lib/core/tree2.js'

// avoid JIT bias
bench('noop', () => {})
bench('noop', () => {})
bench('noop', () => {})
bench('noop', () => {})
bench('noop', () => {})
bench('noop', () => {})

group('built', () => {
  bench('current', () => {
    const tst = new TernarySearchTree()
    for (let i = 0; i < wellknownHeaderNames.length; ++i) {
      const key = wellknownHeaderNames[i].toLowerCase()
      tst.insert(Buffer.from(key), key)
    }
  })
  bench('new', () => {
    const tst = new TernarySearchTree2()
    for (let i = 0; i < wellknownHeaderNames.length; ++i) {
      const key = headerNameLowerCasedRecord[wellknownHeaderNames[i]]
      tst.insert(Buffer.from(key), key)
    }
  })
})

await new Promise((resolve) => setTimeout(resolve, 7000))

await run()
```


```
• built
------------------------------------------------- -----------------------------
current     225.8 µs/iter    (155.4 µs … 1.54 ms)  234.8 µs  654.9 µs  841.7 µs
new          80.6 µs/iter       (62 µs … 2.25 ms)   73.9 µs  262.1 µs  328.4 µs

summary for built
  new
   2.8x faster than current
```